:fire: remove unnecessary layout files
