+++ app/src/main/java/com/example/mobdev21_night_at_the_museum/presentation/intro/IntroductionActivity.java
@@ -3,47 +3,32 @@ package com.example.mobdev21_night_at_the_museum.presentation.intro;
 import android.content.Intent;
 import android.os.Bundle;
 import android.view.View;
-import android.widget.Button;
 
 import androidx.appcompat.app.AppCompatActivity;
 
 import com.example.mobdev21_night_at_the_museum.R;
-import com.google.firebase.auth.FirebaseAuth;
-import com.google.firebase.auth.FirebaseUser;
 
 public class IntroductionActivity extends AppCompatActivity {
-    Button logout;
-    FirebaseUser user;
-    private FirebaseAuth mAuth;
-
+    com.google.android.material.card.MaterialCardView login;
+    com.google.android.material.card.MaterialCardView register;
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.intro1_introduction_activity);
-        mAuth = FirebaseAuth.getInstance();
-        user = mAuth.getCurrentUser();
-        if (user == null) {
-            Intent auth = new Intent(getApplicationContext(), LoginActivity.class);
-            startActivity(auth);
-            finish();
-        }
-        logout = findViewById(R.id.signout);
-
-        logout.setOnClickListener(new View.OnClickListener() {
+        login = findViewById(R.id.loginBtn);
+        register = findViewById(R.id.signupBtn);
+        login.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                signOut();
+                Intent toLogin = new Intent(getApplicationContext(), LoginActivity.class);
+                startActivity(toLogin);
+            }
+        });
+        register.setOnClickListener(new View.OnClickListener(){
+            @Override
+            public void onClick(View view) {
+                Intent toSignUp = new Intent(getApplicationContext(), RegisterActivity.class);
             }
         });
     }
-
-    private void signOut() {
-
-        FirebaseAuth.getInstance().signOut();
-
-        Intent toLogin = new Intent(getApplicationContext(), LoginActivity.class);
-        startActivity(toLogin);
-        finish();
-    }
-
 }
